# docs/api/openapi.yaml
openapi: "3.1.0"
info:
  title: SGG Digital API
  description: |
    API du Secrétariat Général du Gouvernement du Gabon — Plateforme digitale.
    
    ## Modules
    
    | Module | Description |
    |--------|-------------|
    | **Auth** | Authentification JWT (login, register, profil) |
    | **GAR** | Gestion Axée sur les Résultats — PAG 2026 |
    | **Reporting** | Matrice de reporting mensuel des ministères |
    | **Nominations** | Gestion des nominations ministérielles |
    | **Législatif** | Cycle législatif (textes de loi, promulgation) |
    | **e-GOP** | Conseil des Ministres (ordres du jour, PV) |
    | **JO** | Journal Officiel de la République |
    | **PTM** | Programme de Transformation du Monde rural |
    | **Institutions** | Référentiel des institutions gouvernementales |
    | **Monitoring** | Réception des événements de monitoring frontend |
    
    ## Authentification
    
    Toutes les routes protégées nécessitent un JWT Bearer token :
    ```
    Authorization: Bearer <token>
    ```
    
    ## Environnements
    
    | Environnement | URL |
    |---------------|-----|
    | Production | `https://api.sgg.ga` |
    | Staging | `https://staging-api.sgg.ga` |
    | Développement | `http://localhost:8080` |
    
  version: "2.1.0"
  contact:
    name: OKA Tech
    email: dev@okatech.ga
  license:
    name: Propriétaire
    identifier: UNLICENSED
servers:
  - url: https://api.sgg.ga
    description: Production
  - url: https://staging-api.sgg.ga
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT obtenu via POST /api/auth/login
  schemas:
...
    RapportMensuel:
      type: object
      properties:
        id:
          type: string
          format: uuid
        programme_id:
          type: string
          format: uuid
        ministere_id:
          type: string
          format: uuid
        annee:
          type: integer
          example: 2026
        mois:
          type: integer
          example: 1
        statut_validation:
          type: string
          enum: [brouillon, soumis, valide_sgg, valide_sgpr, rejete]
        activites_realisees:
          type: string
        budget_md_fcfa:
          type: number
        engage_md_fcfa:
          type: number
        decaisse_md_fcfa:
          type: number
        pct_avancement_physique:
          type: number
        observations_contraintes:
          type: string
        soumis_par_id:
          type: string
          format: uuid
        valide_sgg_par_id:
          type: string
          format: uuid
        valide_sgpr_par_id:
          type: string
          format: uuid
    ReportingStats:
      type: object
      properties:
        total_rapports:
          type: integer
        brouillons:
          type: integer
        soumis:
          type: integer
        valides_sgg:
          type: integer
        valides_sgpr:
          type: integer
        rejetes:
          type: integer
        budget_total:
          type: number
        moy_execution_financiere:
          type: number

    # ── Monitoring ────────────────────────────────────────────────────────
...
paths:
...
  /api/reporting/matrice:
    get:
      tags: [Reporting]
      summary: Matrice complète de reporting
      parameters:
        - name: annee
          in: query
          schema:
            type: integer
        - name: mois
          in: query
          schema:
            type: integer
        - name: ministere_id
          in: query
          schema:
            type: string
            format: uuid
        - name: statut
          in: query
          schema:
            type: string
            enum: [brouillon, soumis, valide_sgg, valide_sgpr, rejete]
      responses:
        '200':
          description: Matrice avec données enrichies
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/RapportMensuel'
  /api/reporting/stats:
    get:
      tags: [Reporting]
      summary: Statistiques globales de reporting
      parameters:
        - name: annee
          in: query
          schema:
            type: integer
        - name: mois
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Statistiques agrégées
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/ReportingStats'
  /api/reporting/rapports/{id}/submit:
    post:
      tags: [Reporting]
      summary: Soumettre un rapport
      description: "Change le statut de 'brouillon' à 'soumis'. Requiert permission `reporting:write`."
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                activites_realisees:
                  type: string
                budget_md_fcfa:
                  type: number
                pct_avancement_physique:
                  type: number
      responses:
        '200':
          description: Rapport soumis
        '400':
          description: Le rapport ne peut pas être soumis depuis son état actuel
  /api/reporting/rapports/{id}/draft:
    put:
      tags: [Reporting]
      summary: Sauvegarder un brouillon
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Brouillon enregistré
        '404':
          description: Rapport non trouvé
  /api/reporting/rapports/{id}/validate-sgg:
    patch:
      tags: [Reporting]
      summary: Valider un rapport (étape SGG)
      description: "Rôles autorisés : `admin_sgg`, `directeur_sgg`"
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                commentaire:
                  type: string
      responses:
        '200':
          description: Rapport validé par le SGG
        '400':
          description: Le rapport doit être en statut 'soumis'
  /api/reporting/rapports/{id}/validate-sgpr:
    patch:
      tags: [Reporting]
      summary: Valider et publier (étape SGPR)
      description: "Rôles autorisés : `admin_sgg`, `sgpr`, `premier_ministre`"
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                commentaire:
                  type: string
      responses:
        '200':
          description: Rapport validé et publié par le SGPR
  /api/reporting/rapports/{id}/reject:
    patch:
      tags: [Reporting]
      summary: Rejeter un rapport
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [motif]
              properties:
                motif:
                  type: string
                  description: Motif obligatoire du rejet
      responses:
        '200':
          description: Rapport rejeté
        '400':
          description: Motif manquant ou état invalide
  /api/reporting/rapports/batch-validate-sgg:
    post:
      tags: [Reporting]
      summary: Validation en masse (SGG)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [rapportIds]
              properties:
                rapportIds:
                  type: array
                  items:
                    type: string
                    format: uuid
      responses:
        '200':
          description: Nombre de rapports validés
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  count:
                    type: integer
  /api/reporting/suivi:
    get:
      tags: [Reporting]
      summary: Suivi du remplissage par ministère
      parameters:
        - name: annee
          in: query
          schema:
            type: integer
        - name: mois
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: État de remplissage par ministère

  # ── Monitoring ──────────────────────────────────────────────────────────
...